# replit-source/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /** @maxLength 180 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}

export interface ReplyInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}

export interface ConversationReply {
  id: string;
  body: string;
  createdAt: string;
}

export interface CreatedConversation {
  id: string;
  token: string;
}

export interface VisitorConversation {
  name: string;
  body: string;
  createdAt: string;
  replies: ConversationReply[];
}

export type OwnerConversation = VisitorConversation & {
  id: string;
  email: string;
};

export interface HealthStatus {
  status: string;
}

export interface Photo {
  /**
     * @minimum 1
     * @maximum 9
     */
  id: number;
  /**
     * @minLength 1
     * @maxLength 8000000
     */
  src: string;
  /** @maxLength 80 */
  caption: string;
  /** @maxLength 500 */
  description: string;
}

export interface AlbumInput {
  /**
     * @minLength 1
     * @maxLength 48
     */
  title: string;
  /**
     * @minItems 9
     * @maxItems 9
     */
  photos: Photo[];
}

export type AlbumState = AlbumInput & ({
  /** @nullable */
  lastSyncedAt: string | null;
  canEdit: boolean;
  ownerClaimed: boolean;
});