# replit-source/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: album
    description: WD Photo album synchronization
  - name: messages
    description: Private portfolio conversations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /album:
    get:
      operationId: getAlbum
      tags: [album]
      summary: Get the saved photo album
      description: Returns the latest album saved in the connected GitHub repository.
      responses:
        "200":
          description: Current album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumState"
        "502":
          description: GitHub could not be reached
    put:
      operationId: syncAlbum
      tags: [album]
      summary: Save the photo album to GitHub
      description: Saves album metadata and uploaded images to the connected GitHub repository. The first signed-in editor becomes the album owner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumInput"
      responses:
        "200":
          description: Album saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlbumState"
        "400":
          description: Invalid album data
        "401":
          description: Sign-in required
        "403":
          description: Only the album owner can save
        "502":
          description: GitHub could not be updated
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List conversations for the album owner
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnerConversation"
        "401":
          description: Sign-in required
        "403":
          description: Owner access required
    post:
      operationId: createMessage
      tags: [messages]
      summary: Start a private conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedConversation"
        "400":
          description: Invalid input
  /messages/{id}/replies:
    post:
      operationId: replyToMessage
      tags: [messages]
      summary: Reply as the owner
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyInput"
      responses:
        "201":
          description: Reply saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationReply"
        "400":
          description: Invalid reply
        "401":
          description: Sign-in required
        "403":
          description: Owner access required
        "404":
          description: Conversation not found
  /conversations/{token}:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Read a private conversation by its link token
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string, minLength: 40, maxLength: 100 }
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorConversation"
        "404":
          description: Conversation not found
components:
  schemas:
    MessageInput:
      type: object
      required: [name, email, body]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        email: { type: string, format: email, maxLength: 180 }
        body: { type: string, minLength: 1, maxLength: 2000 }
    ReplyInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 2000 }
    ConversationReply:
      type: object
      required: [id, body, createdAt]
      properties:
        id: { type: string, format: uuid }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    CreatedConversation:
      type: object
      required: [id, token]
      properties:
        id: { type: string, format: uuid }
        token: { type: string }
    VisitorConversation:
      type: object
      required: [name, body, createdAt, replies]
      properties:
        name: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        replies:
          type: array
          items: { $ref: "#/components/schemas/ConversationReply" }
    OwnerConversation:
      allOf:
        - $ref: "#/components/schemas/VisitorConversation"
        - type: object
          required: [id, email]
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Photo:
      type: object
      properties:
        id:
          type: integer
          minimum: 1
          maximum: 9
        src:
          type: string
          minLength: 1
          maxLength: 8000000
        caption:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 500
      required:
        - id
        - src
        - caption
        - description
    AlbumInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 48
        photos:
          type: array
          minItems: 9
          maxItems: 9
          items:
            $ref: "#/components/schemas/Photo"
      required:
        - title
        - photos
    AlbumState:
      allOf:
        - $ref: "#/components/schemas/AlbumInput"
        - type: object
          properties:
            lastSyncedAt:
              type: ["string", "null"]
              format: date-time
            canEdit:
              type: boolean
            ownerClaimed:
              type: boolean
          required:
            - lastSyncedAt
            - canEdit
            - ownerClaimed