# replit-source/lib/api-zod/src/generated/types/conversationReply.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConversationReply {
  id: string;
  body: string;
  createdAt: Date;
}