# replit-source/lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /** @maxLength 180 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}