# replit-source/lib/api-zod/src/generated/types/ownerConversation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VisitorConversation } from './visitorConversation';

export type OwnerConversation = VisitorConversation & {
  id: string;
  email: string;
};