# replit-source/lib/api-zod/src/generated/types/replyInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReplyInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}