# replit-source/lib/api-zod/src/generated/types/visitorConversation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationReply } from './conversationReply';

export interface VisitorConversation {
  name: string;
  body: string;
  createdAt: Date;
  replies: ConversationReply[];
}